Backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
pymongo==4.6.1
Flask-PyMongo==2.3.0
PyJWT==2.8.0
Werkzeug==3.0.1
bcrypt==4.1.2
email-validator==2.1.0
groq==0.4.1
